Remove custom metadata builder
